+++ hdbconnect/tests/test_027_small_decimals.rs
@@ -26,128 +26,79 @@ fn test_small_decimals(_log_handle: &mut LoggerHandle, connection: &Connection) 
     connection.multiple_statements_ignore_err(vec!["drop table TEST_SMALL_DECIMALS"]);
 
     let stmts = vec![
-        "create table TEST_SMALL_DECIMALS (f1 NVARCHAR(100) primary key, f2 SMALLDECIMAL, \
-        f3 integer, f2_NN SMALLDECIMAL NOT NULL, f3_NN integer NOT NULL)",
-        "insert into TEST_SMALL_DECIMALS (f1, f2, f2_NN, f3_NN) values('0.00000', 0.000, 0.000, 10)",
-        "insert into TEST_SMALL_DECIMALS (f1, f2, f2_NN, f3_NN) values('0.00100', 0.001, 0.001, 10)",
-        "insert into TEST_SMALL_DECIMALS (f1, f2, f2_NN, f3_NN) values('-0.00100', -0.001, -0.001, 10)",
-        "insert into TEST_SMALL_DECIMALS (f1, f2, f2_NN, f3_NN) values('0.00300', 0.003, 0.003, 10)",
-        "insert into TEST_SMALL_DECIMALS (f1, f2, f2_NN, f3_NN) values('0.00700', 0.007, 0.007, 10)",
-        "insert into TEST_SMALL_DECIMALS (f1, f2, f2_NN, f3_NN) values('0.25500', 0.255, 0.255, 10)",
-        "insert into TEST_SMALL_DECIMALS (f1, f2, f2_NN, f3_NN) values('65.53500', 65.535, 65.535, 10)",
-        "insert into TEST_SMALL_DECIMALS (f1, f2, f2_NN, f3_NN) values('-65.53500', -65.535, -65.535, 10)",
+        "create table TEST_SMALL_DECIMALS (s NVARCHAR(100) primary key, sdec SMALLDECIMAL)",
+        "insert into TEST_SMALL_DECIMALS (s, sdec) values('0.00000', 0.000)",
+        "insert into TEST_SMALL_DECIMALS (s, sdec) values('0.00100', 0.001)",
+        "insert into TEST_SMALL_DECIMALS (s, sdec) values('-0.00100', -0.001)",
+        "insert into TEST_SMALL_DECIMALS (s, sdec) values('0.00300', 0.003)",
+        "insert into TEST_SMALL_DECIMALS (s, sdec) values('0.00700', 0.007)",
+        "insert into TEST_SMALL_DECIMALS (s, sdec) values('0.25500', 0.255)",
+        "insert into TEST_SMALL_DECIMALS (s, sdec) values('65.53500', 65.535)",
+        "insert into TEST_SMALL_DECIMALS (s, sdec) values('-65.53500', -65.535)",
     ];
     connection.multiple_statements(stmts)?;
 
     #[derive(Deserialize)]
     struct TestData {
-        #[serde(rename = "F1")]
-        f1: String,
-        #[serde(rename = "F2")]
-        f2: BigDecimal,
-        #[serde(rename = "F2_NN")]
-        f2_nn: BigDecimal,
-        #[serde(rename = "F3_NN")]
-        f3_nn: u32,
+        #[serde(rename = "S")]
+        s: String,
+        #[serde(rename = "SDEC")]
+        sdec: BigDecimal,
     }
 
-    let insert_stmt_str =
-        "insert into TEST_SMALL_DECIMALS (F1, F2, F2_NN, F3_NN) values(?, ?, ?, ?)";
+    let insert_stmt_str = "insert into TEST_SMALL_DECIMALS (s, sdec) values(?, ?)";
 
     // prepare & execute
-
     let mut insert_stmt = connection.prepare(insert_stmt_str)?;
-    insert_stmt.add_batch(&(
-        "75.53500",
-        BigDecimal::from_f32(75.535).unwrap(),
-        BigDecimal::from_f32(75.535).unwrap(),
-        10,
-    ))?;
-    insert_stmt.add_batch(&("87.65434", 87.654_34_f32, 87.654_34_f32, 10))?;
-    insert_stmt.add_batch(&("0.00500", 0.005001_f32, 0.005001_f32, 10))?;
-    insert_stmt.add_batch(&("-0.00600", -0.006_00_f64, -0.006_00_f64, 10))?;
-    insert_stmt.add_batch(&("-7.65432", -7.654_32_f64, -7.654_32_f64, 10))?;
-    insert_stmt.add_batch(&("99.00000", 99, 99, 10))?;
-    insert_stmt.add_batch(&("-50.00000", -50_i16, -50_i16, 10))?;
-    insert_stmt.add_batch(&("22.00000", 22_i64, 22_i64, 10))?;
+    insert_stmt.add_batch(&("75.53500", BigDecimal::from_f32(75.535).unwrap()))?;
+    insert_stmt.add_batch(&("87.65434", 87.654_34_f32))?;
+    insert_stmt.add_batch(&("0.00500", 0.005001_f32))?;
+    insert_stmt.add_batch(&("-0.00600", -0.006_00_f64))?;
+    insert_stmt.add_batch(&("-7.65432", -7.654_32_f64))?;
+    insert_stmt.add_batch(&("99.00000", 99))?;
+    insert_stmt.add_batch(&("-50.00000", -50_i16))?;
+    insert_stmt.add_batch(&("22.00000", 22_i64))?;
     insert_stmt.execute_batch()?;
 
-    insert_stmt.add_batch(&("-0.05600", "-0.05600", "-0.05600", "10"))?;
-    insert_stmt.add_batch(&("-8.65432", "-8.65432", "-8.65432", "10"))?;
+    insert_stmt.add_batch(&("-0.05600", "-0.05600"))?;
+    insert_stmt.add_batch(&("-8.65432", "-8.65432"))?;
     insert_stmt.execute_batch()?;
 
     info!("Read and verify decimals");
-    let resultset =
-        connection.query("select f1, f2, f3_NN, f2_NN from TEST_SMALL_DECIMALS order by f2")?;
+    let resultset = connection.query("select s, sdec from TEST_SMALL_DECIMALS order by sdec")?;
     let precision = resultset.metadata()[1].precision();
     debug!("metadata: {:?}", resultset.metadata());
-    let scale = 5; //resultset.metadata().scale(1)? as usize;
+    let scale = 5;
     for row in resultset {
         let mut row = row?;
         let s: String = row.next_try_into()?;
-        let bd1: BigDecimal = row.next_try_into()?;
-        row.next_value();
-        let bd2: BigDecimal = row.next_try_into()?;
+        let bd1 = row.next_try_into::<BigDecimal>()?.with_scale(scale);
         debug!("precision = {}, scale = {}", precision, scale);
-        assert_eq!(format!("{}", s), format!("{bd1:.scale$}"));
-        assert_eq!(format!("{}", s), format!("{bd2:.scale$}"));
+        assert_eq!(format!("{}", s), format!("{bd1}"));
     }
 
-    info!("Read and verify decimals to struct");
-    let resultset =
-        connection.query("select f1, f2, f3_NN, f2_NN from TEST_SMALL_DECIMALS order by f2")?;
-    let scale = 5; //resultset.metadata().scale(1)? as usize;
+    info!("Read and verify small decimals to struct");
+    let resultset = connection.query("select s, sdec from TEST_SMALL_DECIMALS order by sdec")?;
+    let scale = 5;
     let result: Vec<TestData> = resultset.try_into()?;
     for td in result {
-        debug!("{:?}, {:?}", td.f1, td.f2);
-        assert_eq!(td.f1, format!("{0:.1$}", td.f2, scale));
-        assert_eq!(td.f1, format!("{0:.1$}", td.f2_nn, scale));
-        assert_eq!(td.f3_nn, 10);
+        assert_eq!(td.s, format!("{}", td.sdec.with_scale(scale)));
     }
 
     info!("Read and verify small decimal to single value");
-    let resultset = connection.query("select AVG(F3) from TEST_SMALL_DECIMALS")?;
-    let mydata: Option<BigDecimal> = resultset.try_into()?;
-    assert_eq!(mydata, None);
+    let resultset = connection.query("select AVG(SDEC) from TEST_SMALL_DECIMALS")?;
+    let _mybigdec: BigDecimal = resultset.try_into().unwrap();
 
-    // info!("Read and verify small decimal to single value");
-    let resultset = connection.query("select AVG(F3_NN) from TEST_SMALL_DECIMALS")?;
-    let mydata: BigDecimal = resultset.try_into()?;
-    assert_eq!(mydata, BigDecimal::from_f32(10.0).unwrap());
-
-    let mydata: Option<i64> = connection
-        .query("select AVG(F2) from TEST_SMALL_DECIMALS where f2 = '65.53500'")?
+    let myi64: i64 = connection
+        .query("select AVG(sdec) from TEST_SMALL_DECIMALS where sdec = '65.53500'")?
         .try_into()?;
-    assert_eq!(mydata, Some(65));
-
-    let mydata: i64 = connection
-        .query("select AVG(F2_NN) from TEST_SMALL_DECIMALS where f2_NN = '65.53500'")?
-        .try_into()?;
-    assert_eq!(mydata, 65);
+    assert_eq!(myi64, 65);
 
     // test failing conversion
-    let mydata: HdbResult<i8> = connection
-        .query("select SUM(ABS(F2)) from TEST_SMALL_DECIMALS")?
+    let myerr: HdbResult<i8> = connection
+        .query("select SUM(ABS(sdec)) from TEST_SMALL_DECIMALS")?
         .try_into();
-    assert!(mydata.is_err());
-
-    // test failing conversion
-    let mydata: HdbResult<i8> = connection
-        .query("select SUM(ABS(F2_NN)) from TEST_SMALL_DECIMALS")?
-        .try_into();
-    assert!(mydata.is_err());
-
-    // test working conversion
-    let mydata: i64 = connection
-        .query("select SUM(ABS(F2)) from TEST_SMALL_DECIMALS")?
-        .try_into()?;
-    assert_eq!(mydata, 481);
-
-    // test working conversion
-    let mydata: i64 = connection
-        .query("select SUM(ABS(F2_NN)) from TEST_SMALL_DECIMALS")?
-        .try_into()?;
-    assert_eq!(mydata, 481);
+    assert!(myerr.is_err());
 
     Ok(())
 }