+++ hdbconnect/tests/test_090_xa_transactions.rs
@@ -183,11 +183,11 @@ mod a {
             tm.start_transaction().unwrap();
 
             debug!("conn_a inserts {}", j);
-            conn_a.dml(&insert_stmt(j + 1, "a"))?;
-            conn_a.dml(&insert_stmt(j + 2, "a"))?;
+            conn_a.dml(insert_stmt(j + 1, "a"))?;
+            conn_a.dml(insert_stmt(j + 2, "a"))?;
             debug!("conn_b inserts {}", j);
-            conn_b.dml(&insert_stmt(j + 3, "b"))?;
-            conn_b.dml(&insert_stmt(j + 4, "b"))?;
+            conn_b.dml(insert_stmt(j + 3, "b"))?;
+            conn_b.dml(insert_stmt(j + 4, "b"))?;
 
             // verify with neutral conn that nothing is visible (count)
             let count: u32 = conn.query(&count_query)?.try_into()?;
@@ -198,11 +198,11 @@ mod a {
 
             tm.start_transaction().unwrap();
             debug!("conn_a inserts {}", j);
-            conn_a.dml(&insert_stmt(j + 1, "a"))?;
-            conn_a.dml(&insert_stmt(j + 2, "a"))?;
+            conn_a.dml(insert_stmt(j + 1, "a"))?;
+            conn_a.dml(insert_stmt(j + 2, "a"))?;
             debug!("conn_b inserts");
-            conn_b.dml(&insert_stmt(j + 3, "b"))?;
-            conn_b.dml(&insert_stmt(j + 4, "b"))?;
+            conn_b.dml(insert_stmt(j + 3, "b"))?;
+            conn_b.dml(insert_stmt(j + 4, "b"))?;
 
             // verify with neutral conn that nothing is visible (count)
             let count: u32 = conn.query(&count_query)?.try_into()?;