+++ hdbconnect_async/tests/test_020_deser.rs
@@ -263,116 +263,123 @@ async fn deser_all_to_string(
     _log_handle: &mut LoggerHandle,
     connection: &Connection,
 ) -> HdbResult<()> {
-    // NULL to not Option
-    assert!(connection
+    // NULL to String
+    let rs = connection
         .query("SELECT TO_BIGINT(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(rs.metadata()[0].type_id().to_string(), "BIGINT".to_string());
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_BINARY(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
-        .query("SELECT TO_BLOB(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
-        .query("SELECT TO_CLOB(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
-        .query("SELECT TO_DATE(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    // assert_eq!(connection.query("SELECT TO_DATS(NULL) FROM DUMMY")?.try_into::<String>().is_err(), true); // TO_DATS returns 00000
-    assert!(connection
+        .await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "VARBINARY".to_string()
+    );
+
+    let rs = connection.query("SELECT TO_BLOB(NULL) FROM DUMMY").await?;
+    assert_eq!(rs.metadata()[0].type_id().to_string(), "BLOB".to_string());
+
+    let rs = connection.query("SELECT TO_DATE(NULL) FROM DUMMY").await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "DAYDATE".to_string()
+    );
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_DECIMAL(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "DECIMAL".to_string()
+    );
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_DOUBLE(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
-        .query("SELECT TO_INT(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(rs.metadata()[0].type_id().to_string(), "DOUBLE".to_string());
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection.query("SELECT TO_INT(NULL) FROM DUMMY").await?;
+    assert_eq!(rs.metadata()[0].type_id().to_string(), "INT".to_string());
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_INTEGER(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
-        .query("SELECT TO_NCLOB(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(rs.metadata()[0].type_id().to_string(), "INT".to_string());
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection.query("SELECT TO_NCLOB(NULL) FROM DUMMY").await?;
+    assert_eq!(rs.metadata()[0].type_id().to_string(), "NCLOB".to_string());
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_NVARCHAR(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
-        .query("SELECT TO_REAL(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "NVARCHAR".to_string()
+    );
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection.query("SELECT TO_REAL(NULL) FROM DUMMY").await?;
+    assert_eq!(rs.metadata()[0].type_id().to_string(), "REAL".to_string());
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_SECONDDATE(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "SECONDDATE".to_string()
+    );
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_SMALLDECIMAL(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "DECIMAL".to_string()
+    );
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_SMALLINT(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
-        .query("SELECT TO_TIME(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "SMALLINT".to_string()
+    );
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection.query("SELECT TO_TIME(NULL) FROM DUMMY").await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "SECONDTIME".to_string()
+    );
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
+
+    let rs = connection
         .query("SELECT TO_TIMESTAMP(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
-    assert!(connection
+        .await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "LONGDATE".to_string()
+    );
+
+    let rs = connection
         .query("SELECT TO_TINYINT(NULL) FROM DUMMY")
-        .await?
-        .try_into::<String>()
-        .await
-        .is_err());
+        .await?;
+    assert_eq!(
+        rs.metadata()[0].type_id().to_string(),
+        "TINYINT".to_string()
+    );
+    assert_eq!(rs.try_into::<String>().await.unwrap(), "<NULL>".to_string());
 
     // NULL to Option
     connection
@@ -512,7 +519,7 @@ async fn deser_all_to_string(
             .await?
             .try_into::<Option<String>>()
             .await?,
-        Some("10".to_string())
+        Some("1E+1".to_string())
     );
     assert_eq!(
         connection
@@ -576,7 +583,7 @@ async fn deser_all_to_string(
             .await?
             .try_into::<Option<String>>()
             .await?,
-        Some("10".to_string())
+        Some("1E+1".to_string())
     );
     assert_eq!(
         connection
@@ -652,7 +659,7 @@ async fn deser_all_to_string(
             .await?
             .try_into::<String>()
             .await?,
-        "10".to_string()
+        "1E+1".to_string()
     );
     assert_eq!(
         connection
@@ -716,7 +723,7 @@ async fn deser_all_to_string(
             .await?
             .try_into::<String>()
             .await?,
-        "10".to_string()
+        "1E+1".to_string()
     );
     assert_eq!(
         connection